[mod] develop 동기화 후 product repository 테스트 진행
