+++ src/main/kotlin/edu/example/kotlindevelop/domain/member/repository/MemberRepository.kt
@@ -2,7 +2,15 @@ package edu.example.kotlindevelop.domain.member.repository
 
 import edu.example.kotlindevelop.domain.member.entity.Member
 import org.springframework.data.jpa.repository.JpaRepository
+import org.springframework.stereotype.Repository
+import java.util.*
 
-interface MemberRepository : JpaRepository<Member, Long> {
+@Repository
+interface MemberRepository : JpaRepository<Member, Long>, MemberRepositoryCustom {
+    fun findByLoginId(loginId: String): Optional<Member>
 
+    fun findByRefreshToken(refreshToken: String): Optional<Member>
+
+    fun findByEmail(email: String): Optional<Member>
+    fun findByLoginIdAndEmail(loginId: String, email: String): Optional<Member>
 }
