+++ src/main/kotlin/edu/example/kotlindevelop/domain/member/service/MemberService.kt
@@ -1,11 +1,11 @@
-package edu.example.kotlindevelop.member.service
+package edu.example.kotlindevelop.domain.member.service
 
 import edu.example.kotlindevelop.global.jwt.JwtUtil
-import edu.example.kotlindevelop.member.dto.MemberDTO
-import edu.example.kotlindevelop.member.entity.Member
-import edu.example.kotlindevelop.member.exception.MemberException
-import edu.example.kotlindevelop.member.repository.MemberRepository
-import edu.example.kotlindevelop.member.util.PasswordUtil
+import edu.example.kotlindevelop.domain.member.dto.MemberDTO
+import edu.example.kotlindevelop.domain.member.entity.Member
+import edu.example.kotlindevelop.domain.member.exception.MemberException
+import edu.example.kotlindevelop.domain.member.repository.MemberRepository
+import edu.example.kotlindevelop.domain.member.util.PasswordUtil
 import io.jsonwebtoken.Claims
 import io.jsonwebtoken.ExpiredJwtException
 import org.modelmapper.ModelMapper
